Flexbox Update
After further reading I decided that Flexbox would be easier for me to use for the homepage grid.

A lot of trial and error went into this new layout, but I think it was worth it. Using flexbox will allow for more real-world content scenarios, where the excerpts may not be uniform in length.